REVISIT: (#6992) Show a summary of nodes x resources on home page.
We now have a summary tab-set on the home page of the Dashboard, showing each
node and the count of "resources by state" that it contains.  This provides
users a solid summary of the state of their network at this point in time.

NOTE: This is incomplete; it needs further attention on both content and
layout before it can be considered finished.